requirements.txt
PyInstaller==3.4
altgraph==0.16.1
appdirs==1.4.3
future==0.17.1
macholib==1.11
pefile==2019.4.18
pycrypto==2.6.1